requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
datasets==4.4.2
dill==0.4.0
dnspython==2.8.0
fastapi==0.128.0
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huepy==1.2.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.6.1
nltk==3.9.2
numpy==2.3.4
numspy==1.0
packaging==25.0
pandas==2.3.3
propcache==0.4.1
psutil==7.2.1
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.15.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.6
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.40.0
xxhash==3.6.0
yarl==1.22.0
